requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
asyncio==3.4.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.2.0
fonttools==4.55.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
httpagentparser==1.9.5
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
ruff==0.9.3
scikit-learn==1.5.0
scipy==1.15.1
shap==0.46.0
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.22
stack-data==0.6.3
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
wcwidth==0.2.13
